Move VideoService to room package
